// SETPPBO/ClientApp/app/main/administrasi-banding-gugatan/models/data-pemohon.ts
export interface DataPemohon {
    PemohonId: string;
    Nama: string;
    NPWP: string;
    Alamat: string;
    Kota: string;
    RefKotaId: number;
    AlamatKoresponden: string;
    KotaKoresponden: string;
    RefKotaKorespondenId: number;
    KodePos: string;
    KodePosKoresponden: string;
}