// SETPPBO/ClientApp/app/main/administrasi-banding-gugatan/models/permintaan-sub-st.ts
export interface PermintaanSubSt {
    PermohonanId: string;
    NoSuratPermintaanSubSt: string;
    TglSuratPermintaanSubSt: Date;
    RefTermohonId: number;
    NamaTermohon: string;
    RefUpTermohonId: number;
    NamaUpTermohon: string;
    AlamatTermohon: string;
    KotaTermohon: string;
    RefTtdPermintaanSubStID: number;
    RefStatusID: number;
    
}